+++ CHANGES.rst
@@ -1,7 +1,7 @@
 Changelog
 =========
 
-1.2 (unreleased)
+1.2 (2020-05-30)
 ----------------
 
 - Add ``generix_required_extractor`` to cron widget extractors.